gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "Gor903",email = "gor.beglaryan.rw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
